Combining migration scripts into one, removing redundant classes, minor renaming and imports notation
