+++ src/main/java/org/ohdsi/webapi/tool/ToolController.java
@@ -3,7 +3,14 @@ package org.ohdsi.webapi.tool;
 import org.ohdsi.webapi.tool.dto.ToolDTO;
 import org.springframework.stereotype.Controller;
 
-import javax.ws.rs.*;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import java.util.List;
