--- src/main/resources/db/migration/postgresql/V2.14.0.20240506100635__add_delete_tool_permission_if_not_exists.sql
@@ -1,18 +1,0 @@
-INSERT INTO  ${ohdsiSchema}.sec_permission (id, value, description)
-SELECT nextval('${ohdsiSchema}.sec_permission_id_seq'), 'tool:*:delete', 'Delete Tool'
-WHERE NOT EXISTS (
-        SELECT NULL FROM  ${ohdsiSchema}.sec_permission
-        WHERE value = 'tool:*:delete'
-);
-
-INSERT INTO ${ohdsiSchema}.sec_role_permission(id, role_id, permission_id)
-SELECT nextval('${ohdsiSchema}.sec_role_permission_sequence'), sr.id, sp.id
-FROM ${ohdsiSchema}.sec_permission SP, ${ohdsiSchema}.sec_role sr
-WHERE sp.value IN (
-    'tool:*:delete'
-    ) AND sr.name IN ('admin')
-  AND NOT EXISTS (
-        SELECT NULL FROM ${ohdsiSchema}.sec_role_permission
-        WHERE permission_id = sp.id and role_id = sr.id);
-
-
